PACK backend fixes
These are based on tests against [PACK:19713] in Skyrim.esm. With the changes in this commit, we now load that form correctly.

Corrected the enum width for PLDT and PTDA's types.

Corrected the control flow for opening subrecords versus reading the current subrecord.

Corrected unflattening of the procedure tree, w.r.t. detecting orphans.

+++ native/src/DovahKit/dovah/data/packages/location_type.h
@@ -2,7 +2,7 @@
 #include <cstdint>
 
 namespace dovah::packages {
-   enum class location_type : uint8_t {
+   enum class location_type : uint32_t {
       reference                   =  0, // REFR form (warns if not persistent)
       interior_cell               =  1, // CELL form
       near_package_start_location =  2,
